Revert "Issue4"
This reverts commit 25bb82cdd641912e8d5526346de7771d42151740.
