+++ app/assets/stylesheets/calendars.scss
@@ -1,4 +1,4 @@
-// Place all the styles related to the plans controller here.
+// Place all the styles related to the calendars controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: https://sass-lang.com/
 body {
