+++ app/helpers/calendars_helper.rb
@@ -1,0 +1,2 @@
+module CalendarsHelper
+end
