--- app/helpers/plans_helper.rb
@@ -1,2 +1,0 @@
-module PlansHelper
-end
