+++ app/views/calendars/index.html.erb
@@ -1,5 +1,5 @@
 <div>
-  <%= form_with model: @plan, url: plans_path, class: 'form' do |f| %>
+  <%= form_with model: @plan, url: calendars_path, class: 'form' do |f| %>
     <%= f.label :日付を選択, class: 'lavel' %>
     <%= f.date_field :date, class: 'date-select' %>
     <%= f.label :予定を入力, class: 'lavel' %>
