+++ config/routes.rb
@@ -1,5 +1,5 @@
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
-  root 'plans#index'
-  resources :plans, only: [:index, :create]
+  root 'calendars#index'
+  resources :calendars
 end
