+++ test/controllers/calendars_controller_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class PlansControllerTest < ActionDispatch::IntegrationTest
+class CalendarsControllerTest < ActionDispatch::IntegrationTest
   # test "the truth" do
   #   assert true
   # end
